# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Club deportivo API",
        "description": "API Reservas club deportivo, es  una pureba tecnica para la empresa NETSNITS ",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Reservas"
        },
        {
            "url": "http://localhost:8000",
            "description": "Class DeportesController"
        },
        {
            "url": "http://localhost:8000",
            "description": "Class PistasController"
        },
        {
            "url": "http://localhost:8000",
            "description": "Class ReservasController"
        },
        {
            "url": "http://localhost:8000",
            "description": "Class SociosController"
        },
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Registrar un nuevo usuario",
                "operationId": "8a56853624e025573120a09a4c75d468",
                "requestBody": {
                    "description": "Datos para registrar un nuevo usuario",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "usuario@dominio.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "contrasena"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "contrasena"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuario registrado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "type": "object"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "token_de_autenticacion"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Iniciar sesión",
                "description": "Inicia sesión y devuelve un token de acceso.",
                "operationId": "a3b306d14572d1f4bd6c064b3233e7b8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "usuario@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "contraseña_secreta"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Inicio de sesión exitoso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Hola usuario@example.com"
                                        },
                                        "access_token": {
                                            "type": "string",
                                            "example": "token_de_acceso_generado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Credenciales no válidas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Detalles de inicio de sesión no válidos"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/logout": {
            "get": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Cerrar sesión",
                "description": "Cierra la sesión del usuario revocando todos los tokens de acceso.",
                "operationId": "3f19b742e4ead13f81ba0d6c0d46ac1b",
                "responses": {
                    "201": {
                        "description": "Cierre de sesión exitoso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Cierre de sesión exitoso"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No se pudo cerrar sesión. Usuario no autenticado."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/perfil": {
            "get": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Listar recursos",
                "description": "Muestra una lista de recursos.",
                "operationId": "4b21a78200b8ed460f84c66d4e7e0069",
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "description": "Información del usuario autenticado",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No autorizado para acceder a este recurso."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/update": {
            "put": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Actualizar recurso",
                "description": "Actualiza el recurso especificado en el almacenamiento.",
                "operationId": "0533045b97c06136dcd13671333c81d4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "nuevoemail@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "contraseña_actual"
                                    },
                                    "newPassword": {
                                        "type": "string",
                                        "example": "nueva_contraseña"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Contraseña actualizada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Contraseña actualizada exitosamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Contraseña actual incorrecta"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "El formato de correo electrónico no es válido."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/delete": {
            "delete": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Eliminar recurso",
                "description": "Elimina el recurso especificado del almacenamiento.",
                "operationId": "585da9399b44d61faae6f2e0f6b76e7d",
                "responses": {
                    "201": {
                        "description": "Usuario eliminado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Usuario eliminado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "No se pudo eliminar el usuario. Usuario no autenticado."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}